requirements.txt
asgiref==3.8.1
Django==5.1.3
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Markdown==3.7
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.2
